feat(dashboard): only provide sql setting for queries using sql
